refactor: Moved involvement of dancers into new class `DancerInvolvement`
